:sparkles: (config) move header config and ccreate getter
