+++ frontend/config/header.json
@@ -1,0 +1,9 @@
+{
+  "logo": "",
+  "structureName": "Parc national des écrins",
+  "menu": {
+    "primaryItemsNumber": 3,
+    "items": [],
+    "supportedLanguage": ["fr"]
+  }
+}
