+++ frontend/src/modules/header/interface.ts
@@ -1,0 +1,16 @@
+interface MenuItem {
+  name: string;
+  url: string;
+}
+
+interface Menu {
+  primaryItemsNumber: number;
+  items: MenuItem[];
+  supportedLanguage: string[];
+}
+
+export interface HeaderConfig {
+  logo: string;
+  structureName: string;
+  menu: Menu;
+}
