+++ frontend/src/modules/header/utills.ts
@@ -1,0 +1,4 @@
+import headerConfig from '../../../config/header.json';
+import { HeaderConfig } from './interface';
+
+export const getHeaderConfig = (): HeaderConfig => headerConfig;
